+++ e2e/commands/init.e2e.1.js
@@ -7,6 +7,7 @@ import { BIT_GIT_DIR, BIT_HIDDEN_DIR, BIT_MAP, BIT_JSON } from '../../src/consta
 import { ScopeJsonNotFound } from '../../src/scope/exceptions';
 import { InvalidBitMap } from '../../src/consumer/bit-map/exceptions';
 import { InvalidBitJson } from '../../src/consumer/bit-json/exceptions';
+import { statusWorkspaceIsCleanMsg } from '../../src/cli/commands/public-cmds/status-cmd';
 
 const assertArrays = require('chai-arrays');
 
@@ -297,7 +298,7 @@ describe('run bit init', function () {
       });
       it('bit status should show nothing-to-tag', () => {
         const output = helper.runCmd('bit status');
-        expect(output).to.have.string('nothing to tag or export');
+        expect(output).to.have.string(statusWorkspaceIsCleanMsg);
       });
     });
   });