+++ e2e/fixtures/compilers/webpack/dev.config.js
@@ -1,0 +1,24 @@
+const merge = require('webpack-merge');
+const baseConfig = require('./base.config.js');
+
+module.exports = merge(baseConfig, {
+  devtool: 'eval-source-map',
+
+  devServer: {
+    inline: true,
+    contentBase: 'src',
+    port: '3001',
+  },
+
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        use: [
+          'style-loader',
+          'css-loader?importLoaders=1',
+        ],
+      },
+    ],
+  },
+});