+++ src/cli/templates/component-issues-template.js
@@ -23,6 +23,8 @@ export function getInvalidComponentLabel(error: Error) {
     case 'MissingFilesFromComponent':
     case 'ComponentNotFoundInPath':
       return 'component files were deleted (use "bit remove [component_id]" to remove the component from your workspace)';
+    case 'ExtensionFileNotFound': // $FlowFixMe error.path is set for ExtensionFileNotFound
+      return `extension file is missing at ${chalk.bold(error.path)}`;
     default:
       return error.name;
   }