+++ src/consumer/bit-json/bit-json.js
@@ -14,6 +14,8 @@ export type BitJsonProps = {
   tester?: string | Testers,
   dependencies?: Object,
   devDependencies?: Object,
+  compilerDependencies?: Object,
+  testerDependencies?: Object,
   packageDependencies?: Object,
   devPackageDependencies?: Object,
   peerPackageDependencies?: Object,
@@ -30,6 +32,8 @@ export default class ComponentBitJson extends AbstractBitJson {
     tester,
     dependencies,
     devDependencies,
+    compilerDependencies,
+    testerDependencies,
     packageDependencies,
     devPackageDependencies,
     peerPackageDependencies,
@@ -37,7 +41,17 @@ export default class ComponentBitJson extends AbstractBitJson {
     bindingPrefix,
     extensions
   }: BitJsonProps) {
-    super({ compiler, tester, dependencies, devDependencies, lang, bindingPrefix, extensions });
+    super({
+      compiler,
+      tester,
+      dependencies,
+      devDependencies,
+      compilerDependencies,
+      testerDependencies,
+      lang,
+      bindingPrefix,
+      extensions
+    });
     this.packageDependencies = packageDependencies || {};
     this.devPackageDependencies = devPackageDependencies || {};
     this.peerPackageDependencies = peerPackageDependencies || {};