+++ src/consumer/versions-ops/checkout-version.js
@@ -173,6 +173,8 @@ async function applyVersion(
   if (componentMap.origin === COMPONENT_ORIGINS.AUTHORED && !id.scope) {
     componentWithDependencies.dependencies = [];
     componentWithDependencies.devDependencies = [];
+    componentWithDependencies.compilerDependencies = [];
+    componentWithDependencies.testerDependencies = [];
     componentWithDependencies.allDependencies = [];
   }
   const rootDir = componentMap.rootDir;