+++ src/scope/component-ops/tag-model-component.js
@@ -38,11 +38,15 @@ function buildComponentsGraph(components: Component[]) {
 
   const graphDeps = new Graph();
   const graphDevDeps = new Graph();
+  const graphCompilerDeps = new Graph();
+  const graphTesterDeps = new Graph();
   components.forEach((component) => {
     setGraphEdges(component, component.dependencies, graphDeps);
     setGraphEdges(component, component.devDependencies, graphDevDeps);
+    setGraphEdges(component, component.compilerDependencies, graphCompilerDeps);
+    setGraphEdges(component, component.testerDependencies, graphTesterDeps);
   });
-  return { graphDeps, graphDevDeps };
+  return { graphDeps, graphDevDeps, graphCompilerDeps, graphTesterDeps };
 }
 
 async function getFlattenedDependencies(
@@ -82,9 +86,8 @@ function updateDependenciesVersions(componentsToTag: Component[]): void {
     const foundDependency = componentsToTag.find(component => component.id.isEqualWithoutVersion(dependency.id));
     if (foundDependency) dependency.id = dependency.id.changeVersion(foundDependency.version);
   };
-  componentsToTag.forEach((componentToTag) => {
-    componentToTag.dependencies.get().forEach(dependency => updateDependencyVersion(dependency));
-    componentToTag.devDependencies.get().forEach(dependency => updateDependencyVersion(dependency));
+  componentsToTag.forEach((oneComponentToTag) => {
+    oneComponentToTag.getAllDependencies().forEach(dependency => updateDependencyVersion(dependency));
   });
 }
 
@@ -252,7 +255,7 @@ export default (async function tagModelComponent({
   // go through all dependencies and update their versions
   updateDependenciesVersions(componentsToTag);
   // build the dependencies graph
-  const { graphDeps, graphDevDeps } = buildComponentsGraph(componentsToTag);
+  const { graphDeps, graphDevDeps, graphCompilerDeps, graphTesterDeps } = buildComponentsGraph(componentsToTag);
 
   const dependenciesCache = {};
   const persistComponent = async (consumerComponent: Component) => {
@@ -301,10 +304,24 @@ export default (async function tagModelComponent({
       graphDevDeps,
       dependenciesCache
     );
+    const flattenedCompilerDependencies = await getFlattenedDependencies(
+      scope,
+      consumerComponent,
+      graphCompilerDeps,
+      dependenciesCache
+    );
+    const flattenedTesterDependencies = await getFlattenedDependencies(
+      scope,
+      consumerComponent,
+      graphTesterDeps,
+      dependenciesCache
+    );
     await scope.sources.addSource({
       source: consumerComponent,
       flattenedDependencies,
       flattenedDevDependencies,
+      flattenedCompilerDependencies,
+      flattenedTesterDependencies,
       message,
       exactVersion,
       releaseType,