+++ src/scope/scope.js
@@ -12,7 +12,7 @@ import ComponentModel from './models/component';
 import { Symlink, Version } from './models';
 import { Remotes } from '../remotes';
 import types from './object-registrar';
-import { propogateUntil, currentDirName, pathHasAll, first, readFile, splitBy, pathNormalizeToLinux } from '../utils';
+import { propogateUntil, currentDirName, pathHasAll, first, splitBy } from '../utils';
 import {
   BIT_HIDDEN_DIR,
   LATEST,
@@ -28,9 +28,7 @@ import { ScopeJson, getPath as getScopeJsonPath } from './scope-json';
 import {
   ScopeNotFound,
   ComponentNotFound,
-  ResolutionException,
   DependencyNotFound,
-  CyclicDependencies,
   MergeConflictOnRemote,
   MergeConflict
 } from './exceptions';
@@ -46,19 +44,12 @@ import VersionDependencies from './version-dependencies';
 import SourcesRepository from './repositories/sources';
 import type { ComponentTree } from './repositories/sources';
 import Consumer from '../consumer/consumer';
-import { index } from '../search/indexer';
 import loader from '../cli/loader';
-import { MigrationResult } from '../migration/migration-helper';
+import type { MigrationResult } from '../migration/migration-helper';
 import migratonManifest from './migrations/scope-migrator-manifest';
 import migrate from './migrations/scope-migrator';
 import type { ScopeMigrationResult } from './migrations/scope-migrator';
-import {
-  BEFORE_PERSISTING_PUT_ON_SCOPE,
-  BEFORE_IMPORT_PUT_ON_SCOPE,
-  BEFORE_MIGRATION,
-  BEFORE_RUNNING_BUILD,
-  BEFORE_RUNNING_SPECS
-} from '../cli/loader/loader-messages';
+import { BEFORE_MIGRATION, BEFORE_RUNNING_BUILD, BEFORE_RUNNING_SPECS } from '../cli/loader/loader-messages';
 import performCIOps from './ci-ops';
 import logger from '../logger/logger';
 import componentResolver from '../component-resolver';
@@ -114,7 +105,7 @@ export default class Scope {
     this.path = scopeProps.path;
     this.scopeJson = scopeProps.scopeJson;
     this.created = scopeProps.created || false;
-    this.tmp = scopeProps.tmp || new Tmp(this);
+    this.tmp = scopeProps.tmp || new Tmp((this: Scope));
     this.sources = scopeProps.sources || new SourcesRepository(this);
     this.objects = scopeProps.objects || new Repository(this, types());
   }
@@ -158,12 +149,19 @@ export default class Scope {
    * @param {BitId} id
    */
   static getComponentRelativePath(id: BitId, scopePath?: string): string {
+    if (!id.scope) {
+      throw new Error('could not find id.scope');
+    }
     const relativePath = pathLib.join(id.name, id.scope);
     if (!id.getVersion().latest) {
+      if (!id.version) {
+        // brought closer because flow can't deduce if it's done in the beginning.
+        throw new Error('could not find id.version');
+      }
       return pathLib.join(relativePath, id.version);
     }
     if (!scopePath) {
-      throw new Error(`could not find the latest version of ${id} without the scope path`);
+      throw new Error(`could not find the latest version of ${id.toString()} without the scope path`);
     }
     const componentFullPath = pathLib.join(scopePath, Scope.getComponentsRelativePath(), relativePath);
     if (!fs.existsSync(componentFullPath)) return '';
@@ -336,7 +334,10 @@ export default class Scope {
       nodePathDirDist &&
       components.some(
         component =>
-          (component.dependencies.isCustomResolvedUsed() || component.devDependencies.isCustomResolvedUsed()) &&
+          (component.dependencies.isCustomResolvedUsed() ||
+            component.devDependencies.isCustomResolvedUsed() ||
+            component.compilerDependencies.isCustomResolvedUsed() ||
+            component.testerDependencies.isCustomResolvedUsed()) &&
           (component.componentMap && component.componentMap.origin === COMPONENT_ORIGINS.AUTHORED) &&
           !component.dists.isEmpty()
       );
@@ -462,6 +463,8 @@ export default class Scope {
         });
         object.flattenedDependencies = getBitIdsWithUpdatedScope(object.flattenedDependencies);
         object.flattenedDevDependencies = getBitIdsWithUpdatedScope(object.flattenedDevDependencies);
+        object.flattenedCompilerDependencies = getBitIdsWithUpdatedScope(object.flattenedCompilerDependencies);
+        object.flattenedTesterDependencies = getBitIdsWithUpdatedScope(object.flattenedTesterDependencies);
         const hashAfter = object.hash().toString();
         if (hashBefore !== hashAfter) {
           logger.debug(`switching ${componentsObjects.component.id()} version hash from ${hashBefore} to ${hashAfter}`);
@@ -529,7 +532,7 @@ export default class Scope {
       'exportManyBareScope',
       'exportManyBareScope: will try to importMany in case there are missing dependencies'
     );
-    const versions = await this.importMany(manyCompVersions.map(compVersion => compVersion.id), undefined, true, false); // resolve dependencies
+    const versions = await this.importMany(manyCompVersions.map(compVersion => compVersion.id), true, false); // resolve dependencies
     logger.debug('exportManyBareScope: successfully ran importMany');
     Analytics.addBreadCrumb('exportManyBareScope', 'exportManyBareScope: successfully ran importMany');
     await this.objects.persist();
@@ -675,8 +678,8 @@ export default class Scope {
     });
   }
 
-  async getObjects(ids: BitIds, withDevDependencies?: boolean): Promise<ComponentObjects[]> {
-    const versions = await this.importMany(ids, withDevDependencies);
+  async getObjects(ids: BitIds): Promise<ComponentObjects[]> {
+    const versions = await this.importMany(ids);
     return Promise.all(versions.map(version => version.toObjects(this.objects)));
   }
 
@@ -694,12 +697,7 @@ export default class Scope {
    * and save them locally.
    * 3. External objects, fetch from a remote and save locally. (done by this.getExternalOnes method).
    */
-  async importMany(
-    ids: BitIds,
-    withEnvironments?: boolean,
-    cache: boolean = true,
-    persist: boolean = true
-  ): Promise<VersionDependencies[]> {
+  async importMany(ids: BitIds, cache: boolean = true, persist: boolean = true): Promise<VersionDependencies[]> {
     logger.debug(`scope.importMany: ${ids.join(', ')}`);
     Analytics.addBreadCrumb('importMany', `scope.importMany: ${Analytics.hashData(ids)}`);
     const idsWithoutNils = removeNils(ids);
@@ -711,7 +709,7 @@ export default class Scope {
     const versionDeps = await Promise.all(
       localDefs.map((def) => {
         if (!def.component) throw new ComponentNotFound(def.id.toString());
-        return def.component.toVersionDependencies(def.id.version, this, def.id.scope, withEnvironments);
+        return def.component.toVersionDependencies(def.id.version, this, def.id.scope);
       })
     );
     logger.debug(
@@ -726,7 +724,7 @@ export default class Scope {
     return versionDeps.concat(externalDeps);
   }
 
-  async importManyOnes(ids: BitId[], cache: boolean = true): Promise<ComponentVersion[]> {
+  async importManyOnes(ids: BitIds, cache: boolean = true): Promise<ComponentVersion[]> {
     logger.debug(`scope.importManyOnes. Ids: ${ids.join(', ')}, cache: ${cache.toString()}`);
     Analytics.addBreadCrumb('importManyOnes', `scope.importManyOnes. Ids: ${Analytics.hashData(ids)}`);
 
@@ -757,23 +755,6 @@ export default class Scope {
     );
   }
 
-  import(id: BitId): Promise<VersionDependencies> {
-    if (!id.isLocal(this.name)) {
-      return this.remotes().then(remotes => this.getExternal({ id, remotes, localFetch: true }));
-    }
-
-    return this.sources.get(id).then((component) => {
-      if (!component) throw new ComponentNotFound(id.toString());
-      return component.toVersionDependencies(id.version, this, this.name);
-    });
-  }
-
-  async get(id: BitId): Promise<ComponentWithDependencies> {
-    return this.import(id).then((versionDependencies) => {
-      return versionDependencies.toConsumer(this.objects);
-    });
-  }
-
   /**
    * return a component only when it's stored locally. Don't go to any remote server and don't throw an exception if the
    * component is not there.
@@ -799,7 +780,7 @@ export default class Scope {
 
     const idsWithoutNils = removeNils(ids);
     if (R.isEmpty(idsWithoutNils)) return Promise.resolve([]);
-    return this.importMany(idsWithoutNils, false, cache).then((versionDependenciesArr: VersionDependencies[]) => {
+    return this.importMany(idsWithoutNils, cache).then((versionDependenciesArr: VersionDependencies[]) => {
       return Promise.all(
         versionDependenciesArr.map(versionDependencies => versionDependencies.toConsumer(this.objects))
       );
@@ -814,7 +795,7 @@ export default class Scope {
     Analytics.addBreadCrumb('getManyWithAllVersions', `scope.getManyWithAllVersions, Ids: ${Analytics.hashData(ids)}`);
     const idsWithoutNils = removeNils(ids);
     if (R.isEmpty(idsWithoutNils)) return Promise.resolve([]);
-    const versionDependenciesArr: VersionDependencies[] = await this.importMany(idsWithoutNils, false, cache);
+    const versionDependenciesArr: VersionDependencies[] = await this.importMany(idsWithoutNils, cache);
 
     const allVersionsP = versionDependenciesArr.map((versionDependencies) => {
       const versions = versionDependencies.component.component.listVersions();
@@ -823,7 +804,8 @@ export default class Scope {
         const versionId = versionDependencies.component.id;
         return versionId.changeVersion(version);
       });
-      return this.importManyOnes(idsWithAllVersions);
+      const bitIdsWithAllVersions = BitIds.fromArray(idsWithAllVersions.filter(x => x));
+      return this.importManyOnes(bitIdsWithAllVersions);
     });
     await Promise.all(allVersionsP);